Fix clock needles not align properly to the current time, remove time script (unused)
