--- Assets/Scripts/Sc_Times.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 38244b58f715da1479ac0a50b4b2f3f6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
